backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
transformers==4.45.0
torch==2.2.2
pillow==10.3.0
httpx==0.27.0
